// src/app/client/teaminformation/models/crossBoardPosition.ts
export class CrossBoardPosition {
    id: string;
    clientExecutiveGuid: string;
    companyGuid: string;
    companyId: number;
    duns: string;
    otherCurrentBoardPosition: string;
    isBoardRole: boolean;
    currentAuditor: string;
    auditorSinceYear: string;
    yrsOnBoard: number;
    committeesListJSON: string;
    modifyBy: string;
    modifyDate: Date | string;
}