// src/app/client/teaminformation/models/auditFirm.ts
export class AuditFirm {
    id: string;
    displayName: string;
    date: Date;
    companyGuid: string;
    companyId: number;
    isBig4: boolean;
    isBig6: boolean;
    shorisTop15tName: boolean;
    isTop30: boolean;
    mappingPattern: string;
    sortOrder: number;
    modifyBy: string;
    modifyDate: Date;
}